refactor: extract numeric data types to independent module
- move UpperLinkData and MetricsData to _stat_numeric.py
- add NumericData to enable basic operations (e.g. add, sub, comparison...) for both UpperLinkData and MetricsData
- corresponding test added
- related files(statistic.py) are updated
